auth_service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
PyJWT==2.9.0
redis==5.1.0
pydantic==2.8.2
pydantic-settings==2.5.2
